+++ demo/stores/app.js
@@ -36,6 +36,10 @@ let data = ImmutableHelper.parseJSON({
     as: 'info',
     text: 'Pill'
   },
+  pod: {
+    border: true,
+    padding: "medium"
+  },
   rainbow: {
     data: [{
       y: "50",
@@ -88,6 +92,9 @@ let data = ImmutableHelper.parseJSON({
   },
   tabs: {
     tabData: [{}, {}]
+  },
+  tile: {
+    content: "Using a borderless Pod can provide padding inside your tile."
   },
   toast: {
     as: "warning",