+++ demo/views/layout/layout.js
@@ -2,6 +2,7 @@ import React from 'react';
 import RowDemo from './row-demo';
 import TabsDemo from './tabs-demo';
 import PodDemo from './pod-demo';
+import TileDemo from './tile-demo';
 
 class Layout extends React.Component {
   /**
@@ -15,6 +16,7 @@ class Layout extends React.Component {
         <TabsDemo />
         <PodDemo />
         <PodDemo collapsed={ true }/>
+        <TileDemo />
       </div>
     );
   }