+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## New Components
 
 * Sidebar 
+* Tile - an alternative to Pod, with a drop shadow.
+
+## Layout Updates
+
+* Pod component now receives two additional props:
+  * `border` - allows developers to disable border.
+  * `padding` - allows developers to have control over padding size.
 
 ## Improve Dialog