+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.8.0
+
+## New Components
+
+* Tile - an alternative to Pod, with a drop shadow.
+
+## Layout Updates
+
+* Pod component now receives two additional props:
+  * `border` - allows developers to disable border.
+  * `padding` - allows developers to have control over padding size.
+
+## Bug Fixes
+
+* Fixes alignment issue with SplitButton when using anchors.
+
 # 0.7.1
 
 ## Updates